terraform/modules/ia-agent/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langsmith
pydantic
google-cloud-pubsub
google-cloud-firestore
requests